Extract create_game/open_round+continue_draft/session-basics MCP tools (Wave C batch 8)
mcp-server.ts 2604 -> 1722 lines. Three new files:

- mcp-game-create-tools.ts (createGameCreateTools): create_game.
- mcp-round-reopen-tools.ts (createRoundReopenTools): open_round,
  continue_draft -- grouped as near-identical twins (feedback
  validation, moderation, daily quota, RELAY_VERBATIM).
- mcp-session-basics-tools.ts (createSessionBasicsTools): get_brief,
  report_progress, screenshot_upload_url, end -- all narrow
  {resolveAuth, injectChannel} shape, same as earlier batches.

RETIRED_GAME_KEY_REASON moves into mcp-tool-support.ts (same
treatment as PLATFORM_CONNECTOR_ONLY_REASON in batch 6), since both
this batch's extracted tools and resolveAuth/start staying behind
need it.

start remains in mcp-server.ts along with the JSON-RPC dispatch and
HTTP route registration below the tools object -- both stay bound to
route-level state (transportSessions/invalidStartsByIp Maps,
presence-pulse wiring, the nudge tracker) that isn't cheap to extract
without a deliberate shared-module step first.
