+++ eslint-rules/module-boundary-map.mjs
@@ -187,6 +187,9 @@ const FILE_BUCKET = {
   'mcp-source-stage-tools': 'agent-surface',
   'mcp-source-patch-tools': 'agent-surface',
   'mcp-source-submit-tools': 'agent-surface',
+  'mcp-game-create-tools': 'agent-surface',
+  'mcp-round-reopen-tools': 'agent-surface',
+  'mcp-session-basics-tools': 'agent-surface',
   'mcp-server-discovery': 'agent-surface',
   'mcp-ui': 'agent-surface',
   'mcp-oauth-metadata': 'agent-surface',
